refactor: 파일을 동일한 경로에서 require 하도록 수정
